requirements.txt
python-telegram-bot[job-queue]==21.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
